Implement ownership check for show, update, and delete notes
Before this commit the user X can perform operations in notes of user Y. To solve this scenario, before each request on actions (show, edit, update, delete) we are checking the ownership of accessed resource. If ownership is false (note user != user authenticated) we are redirecting the user to index path.